organize the project files to facilitate the team work
