+++ app/src/main/AndroidManifest.xml
@@ -12,15 +12,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SplashActivity">
+        <activity android:name=".Activities.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LoginActivity" />
-        <activity android:name=".WelcomeActivity" />
+        <activity android:name=".Activities.LoginActivity" />
+        <activity android:name=".Activities.WelcomeActivity" />
         <activity android:name=".Home"></activity>
     </application>
