+++ app/src/main/java/com/example/gm7/instaproject/Activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.example.gm7.instaproject;
+package com.example.gm7.instaproject.Activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+
+import com.example.gm7.instaproject.R;
 
 public class LoginActivity extends AppCompatActivity {
     private EditText mEdUsername;
