+++ app/src/main/res/layout/activity_splash.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.gm7.instaproject.SplashActivity">
+    tools:context="com.example.gm7.instaproject.Activities.SplashActivity">
 
     <ImageView
         android:layout_width="120dp"
